Kernel: Adjust includes after file reorganization
