+++ Kernel/CMakeLists.txt
@@ -412,28 +412,19 @@ else()
         ${RPI_SOURCES}
 
         Arch/aarch64/BootPPMParser.cpp
-        Arch/aarch64/CrashHandler.cpp        
-        Arch/aarch64/MainIdRegister.cpp        
-        Arch/aarch64/PageDirectory.cpp
-        Arch/aarch64/Utils.cpp
-
+        Arch/aarch64/CrashHandler.cpp
         Arch/aarch64/Dummy.cpp
-
-        # Preload specific
-        Arch/aarch64/init.cpp
-        Arch/aarch64/PrekernelMMU.cpp
-        Arch/aarch64/PrekernelExceptions.cpp
-        Arch/aarch64/PrekernelCommon.cpp
-
-        # Assembly
-        Arch/aarch64/boot.S
-        Arch/aarch64/Aarch64_asm_utils.S
-        Arch/aarch64/vector_table.S
-        Arch/aarch64/SmapDisabler.cpp
+        Arch/aarch64/MainIdRegister.cpp
+        Arch/aarch64/PageDirectory.cpp
         Arch/aarch64/ScopedCritical.cpp
+        Arch/aarch64/SmapDisabler.cpp
+        Arch/aarch64/Utils.cpp
+        Arch/aarch64/init.cpp
+        Arch/aarch64/vector_table.S
 
-        MiniStdLib.cpp
+        # Files from base Kernel
         Heap/kmalloc.cpp
+        MiniStdLib.cpp
 
         Memory/AddressSpace.cpp
         Memory/AnonymousVMObject.cpp
